Plugins: categorize linux.check_syscall as a malware plugin
